requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.27
stripe==7.11.0
paypalrestsdk==1.13.1
coinbase-commerce==1.0.1
python-dotenv==1.0.0
aiosqlite==0.19.0 